// LLD/irctc-train-ticketing-system/src/enums/CoachType.ts
export enum CoachType {
  FIRST_AC = '1A',           // First AC
  SECOND_AC = '2A',          // Second AC
  THIRD_AC = '3A',           // Third AC
  SLEEPER = 'SL',            // Sleeper
  CHAIR_CAR = 'CC',          // Chair Car
  AC_CHAIR_CAR = 'ACC',      // AC Chair Car
  GENERAL = 'GEN',           // General
  SECOND_SEATING = '2S'      // Second Seating
}